requirements.txt
torch
torchvision
pandas
pillow
spacy
fastapi
uvicorn
tqdm
python-multipart
matplotlib
